requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
SQLAlchemy>=2.0.32
pyodbc>=5.1.0
sqlglot>=25.0.0
openai>=1.45.0
langgraph>=0.2.20
langchain-core>=0.3.13
python-dotenv>=1.0.1
PyYAML>=6.0.2
pandas>=2.2.2
plotly>=5.24.0
orjson>=3.10.7
tenacity>=8.3.0
rich>=13.9.2
chromadb>=0.5.0
sentence-transformers>=2.2.2
sqlparse>=0.5.0
tabulate>=0.9.0